New mp4 extractor

+++ bibliopixel/util/image/gif_writer.py
@@ -1,5 +1,5 @@
 import os, shutil, statistics, tempfile, time, traceback
-from . import gif, renderer
+from . import gif, mp4, renderer
 from .. import colors, log
 
 
@@ -13,29 +13,42 @@ DEFAULT_RENDER = {
     'padding': 1,
 }
 
+WRITERS = {
+    '.gif': gif.write_gif,
+    '.mp4': mp4.write,
+}
+
 
 class GifWriter:
     """Write an animated GIF given frames from an animation."""
 
     def __init__(self, filename='gif_writer', render=None,
                  divide=1, frames=128, time=10, speed=1.0, gif_options=None,
-                 tmp_dir=None):
+                 tmp_dir=None, remove_tmp_dir=False, suffix='.gif',
+                 duration=None, fps=None):
         self.render = dict(DEFAULT_RENDER, **(render or {}))
         self.divide = divide
         self.frames = frames
         self.time = time
         self.frame_files = []
         self.speed = speed
+        self.duration = duration
+        self.fps = fps
         self.gif_options = gif_options or {}
+        try:
+            self.writer = WRITERS[suffix]
+        except:
+            raise ValueError('Cannot write %s files' % suffix)
 
         filename = os.path.expanduser(os.path.abspath(filename))
-        file_root = filename[:-4] if filename.endswith('.gif') else filename
-        self.filename = file_root + '.gif'
+        file_root = filename[:-4] if filename.endswith(suffix) else filename
+        self.filename = file_root + suffix
         os.makedirs(os.path.dirname(self.filename), exist_ok=True)
 
         if tmp_dir:
             self.tmp_dir_name = tmp_dir
-            shutil.rmtree(tmp_dir, ignore_errors=True)
+            if remove_tmp_dir:
+                shutil.rmtree(tmp_dir, ignore_errors=True)
             os.makedirs(self.tmp_dir_name, exist_ok=True)
         else:
             self.tmp_dir = tempfile.TemporaryDirectory()
@@ -67,35 +80,36 @@ class GifWriter:
             self.write()
             return True
 
-        frame_name = '%s%04d.gif' % (self.basename, frame)
+        frame_name = '%s%04d.png' % (self.basename, frame)
         filename = os.path.join(self.tmp_dir_name, frame_name)
 
         self.render().save(filename)
         self.frame_files.append(filename)
 
+    @property
+    def fps(self):
+        if self._fps:
+            print('1')
+            return self._fps
+
+        if self.duration:
+            print('2')
+            return 1 / self.duration
+
+        if len(self.frame_files) < 2:
+            print('3')
+            return 1
+
+        dt = self.times[-1] - self.times[0]
+        slots = len(self.frame_files) - 1
+        print('4')
+        return slots / dt
+
+    @fps.setter
+    def fps(self, fps):
+        self._fps = fps
+
     def write(self):
-        if not self.frame_files:
-            return
-
-        ff, self.frame_files = self.frame_files, []
-        go = dict(self.gif_options)
-
-        if 'duration' in go:
-            go.pop('fps', None)
-            go['duration'] /= self.speed
-            assert isinstance(go['duration'], (int, float))
-
-        elif 'fps' in go:
-            go['fps'] *= self.speed
-
-        else:
-            # Compute the average duration
-            dt = [t1 - t0 for t0, t1 in zip(self.times, self.times[1:])]
-            duration = statistics.mean(dt)
-
-            # GIF durations are only expressed in hundredths of seconds.
-            duration = round(duration, 2)
-            go = dict(go, duration=duration)
-
-        gif.write_gif(self.filename, ff, **go)
-        self.tmp_dir = None
+        fps = self.fps * self.speed
+        print('fps', fps)
+        self.writer(self.filename, self.frame_files, fps, **self.gif_options)
